test(#991): extract parity helpers into sibling module to satisfy file-size ratchet (#1135)
